requirements.txt
fastapi
httpx
requests
uvicorn
python-jose
pyjwt